Get rid of NOCOPY_PRIVATE_FIELD
I'm not really certain how useful this is, and the fact that the
implementation requires passing in the containing class name to the
macro is a bit fugly. I believe the original intention was to prevent
consumers of the reference and archive classes from instantiating tag
classes that they should not have access to by exploiting the fact that
these classes are encoded as a part of the types that consumers *do*
have access to. Unless a stronger use case arises, I think I prefer to
accept that a determined user might intentionally subvert the system and
leave it at that.